Fix 'test' login provider
Since I made the login button into an actual <button> element, login
was broken. Turns out to be because the usual 'submit' behaviour of the
form would happen just after the onclick handler was called, which of
course interrupted the AJAX request before it finished.

That's 3 hours of my life I'll never get back.